+++ pkg/proxy/integrations/mysqlparser/README.md
@@ -1,0 +1,6 @@
+# MySQL Package Documentation
+
+The `mysqlparser` package encompasses the parser and mapping logic required 
+to read MySql binary messages and capture and test the outputs. 
+Utilized by the `hooks` package, it assists in redirecting outgoing 
+calls for the purpose of recording or testing the outputs.
